feat: add auth provider component
