+++ CoShift/frontend/src/app/App.tsx
@@ -1,82 +1,21 @@
-import { useState, useCallback, useEffect } from 'react'
 import './App.css'
 import Login       from '../features/auth/components/LoginForm'
 import WeekPage    from '../pages/WeekPage'
 import Layout      from '../layout/Layout'
 import PrivateLayout from '../layout/PrivateLayout'
-import { AuthContext } from '../features/auth/hooks/AuthContext'
 import { Routes, Route, Navigate } from 'react-router-dom'
 import AdminPage   from '../pages/AdminPage'
+import { AuthProvider } from '../features/auth/AuthProvider'
+import { useAuth } from '../features/auth/hooks/AuthContext'
 
-/**
- * Root-Komponente:
- * – Verwaltet den Auth-Header im State.
- * – Reicht eine „tryLogin“-Funktion an <Login/> durch, die
- *   einen Probe-Call auf ein geschütztes Backend-API macht.
- * – Zeigt nach erfolgreichem Login die <WeekPage/>.
- */
-export default function App() {
-  const [authHeader, setAuthHeader] = useState<string | null>(null)
-  const [balance,    setBalance]    = useState<number | null>(null)
-
-
-
-  const tryLogin = useCallback(async (header: string): Promise<boolean> => {
-    try {
-      const res = await fetch('/api/shifts', {
-        headers: { Authorization: header }
-      })
-
-      if (res.ok) {
-        setAuthHeader(header)
-        sessionStorage.setItem('authHeader', header)
-
-        // Saldo laden
-        try {
-          const balRes = await fetch('/api/persons/me/timeaccount', {
-            headers: { Authorization: header }
-          })
-          if (balRes.ok) {
-            const dto = await balRes.json() as { balanceInMinutes: number }
-            setBalance(dto.balanceInMinutes)
-          }
-        } catch (e) {
-          console.error('Unable to fetch balance', e)
-        }
-
-        return true
-      }
-    } catch {
-      console.error('Api unavailable')
-    }
-    sessionStorage.removeItem('authHeader')
-    return false                         // 401 
-  }, [])
-
-  function logout() {
-    setAuthHeader(null)
-    setBalance(null)
-    sessionStorage.removeItem('authHeader')        
-  }
-
-  useEffect(() => {
-    const stored = sessionStorage.getItem('authHeader')
-    if (stored) {
-      // Schnelltest: ist das Token noch gültig?
-      tryLogin(stored).catch(() => {
-        sessionStorage.removeItem('authHeader')
-      })
-    }
-  }, [tryLogin])
-  
-
-
+function Routing() {
+  const { header, tryLogin } = useAuth()
 
   return (
-    <AuthContext.Provider value={{ header: authHeader, balance, logout }}>
-      {authHeader && <Layout />}          {/* <- MUI-AppBar inkl. Burger */}
+    <>
+      {header && <Layout />}
 
-      {!authHeader ? (
+      {!header ? (
         <Login onLogin={tryLogin} />
       ) : (
         <Routes>
@@ -87,6 +26,14 @@ export default function App() {
           <Route path="*" element={<Navigate to="/" />} />
         </Routes>
       )}
-    </AuthContext.Provider>
+    </>
+  )
+}
+
+export default function App() {
+  return (
+    <AuthProvider>
+      <Routing />
+    </AuthProvider>
   )
 }
