+++ CoShift/frontend/src/features/auth/hooks/AuthContext.tsx
@@ -1,9 +1,9 @@
-import { createContext, useContext } from 'react'
+import { useContext } from 'react'
+import { AuthContext } from '../AuthProvider'
 import type { AuthCtx } from '../types'
 
-export const AuthContext = createContext<AuthCtx | null>(null)
 export const useAuth = () => {
-  const ctx = useContext(AuthContext)
-  if (!ctx) throw new Error('useAuth must be inside <AuthContext.Provider>')
+  const ctx = useContext<AuthCtx | null>(AuthContext)
+  if (!ctx) throw new Error('useAuth must be inside <AuthProvider>')
   return ctx
 }
