+++ CoShift/frontend/src/features/auth/types/index.ts
@@ -1,5 +1,6 @@
 export interface AuthCtx {
   header: string | null
   balance: number | null
+  tryLogin: (authHeader: string) => Promise<boolean>
   logout: () => void
 }
